EX1/Deliverables/requirements.txt
docx-mailmerge==0.5.0
pyyaml==6.0.2